+++ orator/orm/relations/morph_one.py
@@ -32,3 +32,12 @@ class MorphOne(MorphOneOrMany):
         :type relation:  str
         """
         return self.match_one(models, results, relation)
+
+    def new_instance(self, parent):
+        return MorphOne(
+            self._related.new_query(),
+            parent,
+            self._morph_type,
+            self._foreign_key,
+            self._local_key
+        )